handle flags and defaults better. checked uniqer

+++ main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/rotblauer/bildRoam/bildRoam"
 	"io"
 	"os/exec"
+	"os/user"
 	// "log"
 	"encoding/csv"
 	"os"
@@ -22,20 +23,58 @@ func floatToString(input float64) string {
 	return strconv.FormatFloat(input, 'f', 6, 64)
 }
 
+//http://stackoverflow.com/questions/10510691/how-to-check-whether-a-file-or-directory-denoted-by-a-path-exists-in-golang
+// exists returns whether the given file or directory exists or not
+func pathExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return true, err
+}
+
 func main() {
 	var output string
 	var targetDir string
 	var myName string
 	var uniqify bool
 
-	// go run main.go -dir ~/Desktop/locs/ -out ~/Desktop/locs/whois.csv -name jl
+	//get user's home dir so default to iphotos lib
+	usr, err := user.Current()
+	if err != nil {
+		fmt.Println("but whom", err)
+	}
+	hd := usr.HomeDir
+	defaultiPhotosLibPath := "Pictures/Photos Library.photoslibrary/Masters"
+	defaultiPhotosLibPath, e := filepath.Abs(filepath.Join(hd, defaultiPhotosLibPath))
+	if e != nil {
+		fmt.Println("had a hard time setting default iphotos lib path")
+	}
+
+	//set default tag name
+	uname := usr.Username
+	if uname == "" {
+		uname = usr.Name
+	}
+	if uname == "" {
+		uname = filepath.Base(hd)
+	}
 
 	flag.StringVar(&output, "out", "iwazhere.csv", "specify the output .csv file")
-	flag.StringVar(&targetDir, "dir", "/Users/ia/Pictures/Photos Library.photoslibrary/Masters", "directory with photos")
-	flag.StringVar(&myName, "name", "ia", "your tag name")
+	flag.StringVar(&targetDir, "dir", defaultiPhotosLibPath, "directory with photos")
+	flag.StringVar(&myName, "name", uname, "your tag name")
 	flag.BoolVar(&uniqify, "uniq", true, "make the output csv pipe through sort -u after")
 
 	flag.Parse()
+
+	if exists, _ := pathExists(targetDir); !exists {
+		fmt.Println("You said the photos would be here:", targetDir)
+		fmt.Println("And there just ain't no such place.")
+		return
+	}
 
 	file, err := os.Create(output)
 	if err != nil {
@@ -62,9 +101,6 @@ func main() {
 			}
 		}
 
-		// save to... csv?
-		// fmt.Println(path, lat, lng, ti)
-		// counter++
 		return nil //nil
 	})
